Split InvalidPointer out of ds/dllist.h
This lets us use Pagemaps without requiring dllists in scope
